error correction
.DS_Store из репозитория и добавил его в .gitignore
Добавл версию
Исправил работу с API
Испправил модульность
Исправил myId
испраил сборку
Переместил в src
Изменил readme
Добавил версию
