+++ src/blocks/popup/__photo-content/popup__photo-content.css
