+++ src/blocks/popup/__photo/popup__photo.css
