+++ src/script/CardList.js
@@ -1,7 +1,10 @@
 'use strict';
 
-import {Card} from './Card.js'
+import {Card} from './Card.js';
+import {api} from './Api.js';
 //контейнер карточек
+export let cardContainer = '';
+
 export class CardList{
     constructor (domContainer, cardArray) {
       this.domContainer = domContainer;
@@ -21,3 +24,7 @@ export class CardList{
       this.domContainer.appendChild(card);
     }
   }
+
+  api.getCards().then(data => {
+    cardContainer = new CardList(document.querySelector('.places-list'), data);
+  })
