+++ webpack.config.js
@@ -1,8 +1,8 @@
 // webpack.config.js
 const webpack = require('webpack');
-new webpack.DefinePlugin({
-    'NODE_ENV': JSON.stringify(process.env.NODE_ENV)
-})
+// new webpack.DefinePlugin({
+//     'NODE_ENV': JSON.stringify(process.env.NODE_ENV)
+// })
 
 const path = require('path');
 // Подключили к проекту плагин
@@ -17,7 +17,7 @@ const OptimizeCssAssetsPlugin = require('optimize-css-assets-webpack-plugin');
 const isDev = process.env.NODE_ENV === 'development';
 // создаем переменную для development-сборки
 module.exports = {
-    entry: { main: './script/script.js' },
+    entry: { main: './src/script/index.js' },
     output: {
         path: path.resolve(__dirname, 'dist'),
         filename: '[name].[chunkhash].js'
@@ -45,7 +45,7 @@ module.exports = {
         {
             test: /\.(png|jpg|gif|ico|svg)$/,
             use: [
-                 'file-loader?name=../images/[name].[ext]', // указали папку, куда складывать изображения
+                 'file-loader?name=../src/images/[name].[ext]', // указали папку, куда складывать изображения
                  {
                      loader: 'image-webpack-loader',
                      options: {mozjpeg: {
@@ -99,10 +99,13 @@ module.exports = {
         new HtmlWebpackPlugin({
             // Означает, что:
             inject: false, // стили НЕ нужно прописывать внутри тегов
-            template: './index.html', // откуда брать образец для сравнения с текущим видом проекта
+            template: './src/index.html', // откуда брать образец для сравнения с текущим видом проекта
             filename: 'index.html' // имя выходного файла, то есть того, что окажется в папке dist после сборки
           }),
-          new WebpackMd5Hash()
+          new WebpackMd5Hash(),
+          new webpack.DefinePlugin({
+            'NODE_ENV': JSON.stringify(process.env.NODE_ENV)
+        })
     ]
 }
